refactor(pkg/core/xds): remove dependency on pkg/xds/envoy
