+++ pkg/xds/envoy/clusters/v3/client_side_mtls_configurer_test.go
@@ -6,11 +6,11 @@ import (
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	core_mesh "github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
-	core_xds "github.com/kumahq/kuma/pkg/core/xds"
 	test_model "github.com/kumahq/kuma/pkg/test/resources/model"
 	util_proto "github.com/kumahq/kuma/pkg/util/proto"
 	"github.com/kumahq/kuma/pkg/xds/envoy"
 	"github.com/kumahq/kuma/pkg/xds/envoy/clusters"
+	"github.com/kumahq/kuma/pkg/xds/envoy/tags"
 )
 
 var _ = Describe("EdsClusterConfigurer", func() {
@@ -18,7 +18,7 @@ var _ = Describe("EdsClusterConfigurer", func() {
 	type testCase struct {
 		clusterName   string
 		clientService string
-		tags          []envoy.Tags
+		tags          []tags.Tags
 		mesh          *core_mesh.MeshResource
 		expected      string
 	}
@@ -26,7 +26,7 @@ var _ = Describe("EdsClusterConfigurer", func() {
 	DescribeTable("should generate proper Envoy config",
 		func(given testCase) {
 			// when
-			tracker := core_xds.NewSecretsTracker(given.mesh.GetMeta().GetName(), nil)
+			tracker := envoy.NewSecretsTracker(given.mesh.GetMeta().GetName(), nil)
 			cluster, err := clusters.NewClusterBuilder(envoy.APIV3).
 				Configure(clusters.EdsCluster(given.clusterName)).
 				Configure(clusters.ClientSideMTLS(tracker, given.mesh, given.clientService, true, given.tags)).
@@ -111,7 +111,7 @@ var _ = Describe("EdsClusterConfigurer", func() {
 					},
 				},
 			},
-			tags: []envoy.Tags{
+			tags: []tags.Tags{
 				map[string]string{
 					"kuma.io/service": "backend",
 					"cluster":         "1",
@@ -204,7 +204,7 @@ var _ = Describe("EdsClusterConfigurer", func() {
 					},
 				},
 			},
-			tags: []envoy.Tags{
+			tags: []tags.Tags{
 				{
 					"kuma.io/service": "backend",
 					"version":         "v1",
