+++ pkg/xds/envoy/clusters/v3/client_side_tls_configurer.go
@@ -9,8 +9,8 @@ import (
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	"github.com/kumahq/kuma/pkg/core/xds"
 	"github.com/kumahq/kuma/pkg/util/proto"
-	envoy_common "github.com/kumahq/kuma/pkg/xds/envoy"
 	envoy_metadata "github.com/kumahq/kuma/pkg/xds/envoy/metadata/v3"
+	"github.com/kumahq/kuma/pkg/xds/envoy/tags"
 	envoy_tls "github.com/kumahq/kuma/pkg/xds/envoy/tls/v3"
 )
 
@@ -56,7 +56,7 @@ func (c *ClientSideTLSConfigurer) Configure(cluster *envoy_cluster.Cluster) erro
 			cluster.TransportSocketMatches = append(cluster.TransportSocketMatches, &envoy_cluster.Cluster_TransportSocketMatch{
 				Name: ep.Target,
 				Match: &structpb.Struct{
-					Fields: envoy_metadata.MetadataFields(envoy_common.Tags(ep.Tags).WithoutTags(mesh_proto.ServiceTag)),
+					Fields: envoy_metadata.MetadataFields(tags.Tags(ep.Tags).WithoutTags(mesh_proto.ServiceTag)),
 				},
 				TransportSocket: transportSocket,
 			})
