+++ pkg/xds/envoy/listeners/v3/server_mtls_configurer_test.go
@@ -29,7 +29,7 @@ var _ = Describe("ServerMtlsConfigurer", func() {
 	DescribeTable("should generate proper Envoy config",
 		func(given testCase) {
 			// when
-			tracker := core_xds.NewSecretsTracker(given.mesh.GetMeta().GetName(), nil)
+			tracker := envoy_common.NewSecretsTracker(given.mesh.GetMeta().GetName(), nil)
 			listener, err := NewListenerBuilder(envoy_common.APIV3).
 				Configure(InboundListener(given.listenerName, given.listenerAddress, given.listenerPort, given.listenerProtocol)).
 				Configure(FilterChain(NewFilterChainBuilder(envoy_common.APIV3).
