+++ pkg/xds/envoy/tls/sni.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	"github.com/kumahq/kuma/pkg/xds/envoy"
+	"github.com/kumahq/kuma/pkg/xds/envoy/tags"
 )
 
-func SNIFromTags(tags envoy.Tags) string {
+func SNIFromTags(tags tags.Tags) string {
 	extraTags := tags.WithoutTags(mesh_proto.ServiceTag).String()
 	service := tags[mesh_proto.ServiceTag]
 	if extraTags == "" {
