+++ pkg/xds/generator/secrets/generator_test.go
@@ -56,7 +56,7 @@ var _ = Describe("SecretsGenerator", func() {
 						Mesh: "demo",
 					},
 				},
-				SecretsTracker: core_xds.NewSecretsTracker("", nil),
+				SecretsTracker: envoy_common.NewSecretsTracker("", nil),
 				APIVersion:     envoy_common.APIV3,
 			},
 		}),
@@ -87,7 +87,7 @@ var _ = Describe("SecretsGenerator", func() {
 						},
 					},
 				},
-				SecretsTracker: core_xds.NewSecretsTracker("", nil),
+				SecretsTracker: envoy_common.NewSecretsTracker("", nil),
 				APIVersion:     envoy_common.APIV3,
 			},
 		}),
@@ -155,7 +155,7 @@ var _ = Describe("SecretsGenerator", func() {
 						},
 					},
 				},
-				SecretsTracker: core_xds.NewSecretsTracker("default", []string{"default"}),
+				SecretsTracker: envoy_common.NewSecretsTracker("default", []string{"default"}),
 				APIVersion:     envoy_common.APIV3,
 			},
 			identity: true,
@@ -252,7 +252,7 @@ var _ = Describe("SecretsGenerator", func() {
 			},
 			proxy: &core_xds.Proxy{
 				Id:             *core_xds.BuildProxyId("", mesh_proto.ZoneEgressServiceName),
-				SecretsTracker: core_xds.NewSecretsTracker("mesh-1", []string{"mesh-1", "mesh-2"}),
+				SecretsTracker: envoy_common.NewSecretsTracker("mesh-1", []string{"mesh-1", "mesh-2"}),
 				APIVersion:     envoy_common.APIV3,
 			},
 			expected: "envoy-config-dataplane.golden.yaml",
@@ -328,7 +328,7 @@ var _ = Describe("SecretsGenerator", func() {
 						},
 					},
 				},
-				SecretsTracker: core_xds.NewSecretsTracker("mesh-2", []string{"mesh-1", "mesh-2"}),
+				SecretsTracker: envoy_common.NewSecretsTracker("mesh-2", []string{"mesh-1", "mesh-2"}),
 			},
 			identity: true,
 			usedCas: map[string]struct{}{
